--- apps/web/app/edit-template/page.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import Link from "next/link";
-
-import Head from "../../components/main/head";
-
-function Index(props: any) {
-  return (
-    <div>
-      <Head />
-      456
-    </div>
-  );
-}
-
-export default Index;
