+++ apps/web/components/page-components/login/LoginBackGround.tsx
@@ -2,7 +2,7 @@ import { useIsServer } from "@/hooks/useIsServer";
 import { useTheme } from "next-themes";
 import Link from "next/link";
 
-import Planet from "./animation/Planet";
+import Planet from "../../animation/Planet";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   const { setTheme, theme } = useTheme();
@@ -15,8 +15,8 @@ export default function Layout({ children }: { children: React.ReactNode }) {
       <div className="h-[100vh] w-[100vw] flex overflow-hidden">
         <div className="flex-1 bg-gradient-to-r from-red-600 to-red-500 bg-design">
           <div className="h-full flex flex-col justify-between p-5 text-white font-serif">
-            <p className="text-3xl">Detheroc</p>
-            <p className="text-5xl">Lets Kick Now！</p>
+            <p className="text-3xl">POSTERCRAFT</p>
+            <p className="text-5xl">Lets Create Now！</p>
             <Link href="/">&lt; Home</Link>
           </div>
         </div>
